# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/calls:
    get:
      tags:
        - Calls
      summary: List Calls
      description: Get a paginated list of Calls with optional filtering and sorting
      operationId: listCalls
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Calls
      summary: Create Call
      description: Create a new Call
      operationId: createCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCallRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Call'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Call:
      type: object
      required:
        - id
        - direction
        - from_number
        - to_number
        - status
        - started_at
        - duration_seconds
        - metadata
      properties:
        id:
          type: string
          format: uuid
        direction:
          type: string
        from_number:
          type: string
        to_number:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        agent_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        external_id:
          nullable: true
          type: string
        subject_type:
          nullable: true
          type: string
        subject_id:
          nullable: true
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        answered_at:
          nullable: true
          type: string
          format: date-time
        ended_at:
          nullable: true
          type: string
          format: date-time
        duration_seconds:
          type: integer
          format: int32
        recording_url:
          nullable: true
          type: string
        notes:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateCallRequest:
      type: object
      required:
        - direction
        - from_number
        - to_number
        - status
        - started_at
        - duration_seconds
        - metadata
      properties:
        direction:
          type: string
        from_number:
          type: string
        to_number:
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        agent_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        external_id:
          nullable: true
          type: string
        subject_type:
          nullable: true
          type: string
        subject_id:
          nullable: true
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        answered_at:
          nullable: true
          type: string
          format: date-time
        ended_at:
          nullable: true
          type: string
          format: date-time
        duration_seconds:
          type: integer
          format: int32
        recording_url:
          nullable: true
          type: string
        notes:
          nullable: true
          type: string
        metadata:
          type: object
...
    CallList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Call'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []