# src/exports/types.rs
//! Public types for Telephony module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use crate::domain::entity::*;
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// ============================================================================
// CALL TYPES
// ============================================================================

/// Type-safe ID for Call
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct CallId(pub Uuid);

impl CallId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for CallId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<CallId> for Uuid {
    fn from(id: CallId) -> Self {
        id.0
    }
}

/// Data transfer object for Call
///
/// This is the public representation of Call for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CallDto {
    pub id: CallId,
    pub direction: CallDirection,
    pub from_number: String,
    pub to_number: String,
    pub party_id: Option<Uuid>,
    pub agent_id: Option<Uuid>,
    pub status: CallStatus,
    pub external_id: Option<String>,
    pub subject_type: Option<String>,
    pub subject_id: Option<Uuid>,
    pub started_at: DateTime<Utc>,
    pub answered_at: Option<DateTime<Utc>>,
    pub ended_at: Option<DateTime<Utc>>,
    pub duration_seconds: i32,
    pub recording_url: Option<String>,
    pub notes: Option<String>,
    pub metadata: serde_json::Value,
}

/// Summary view of Call for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CallSummary {
    pub id: CallId,
    pub status: CallStatus,
}

/// Reference to Call for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CallRef {
    pub id: CallId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>